+++ src/main/java/chess/domain/command/Start.java
@@ -3,7 +3,6 @@ package chess.domain.command;
 import chess.domain.board.Position;
 import chess.domain.piece.Piece;
 import chess.domain.state.State;
-
 import java.util.Map;
 
 public class Start extends Command {
