+++ src/main/java/chess/domain/command/Status.java
@@ -3,7 +3,6 @@ package chess.domain.command;
 import chess.domain.board.Result;
 import chess.domain.piece.Color;
 import chess.domain.state.State;
-
 import java.util.Map;
 
 public class Status extends Start {
