+++ src/main/java/chess/domain/piece/Pawn.java
@@ -2,6 +2,7 @@ package chess.domain.piece;
 
 import chess.domain.board.Direction;
 import chess.domain.board.Position;
+import chess.domain.piece.movePattern.AbstractPawnMovePattern;
 
 public final class Pawn extends Piece {
