+++ src/main/java/chess/domain/piece/Piece.java
@@ -2,7 +2,6 @@ package chess.domain.piece;
 
 import chess.domain.board.Direction;
 import chess.domain.board.Position;
-
 import java.util.Objects;
 
 public abstract class Piece {
@@ -33,8 +32,12 @@ public abstract class Piece {
 
     @Override
     public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
         Piece piece = (Piece) o;
         return color == piece.color;
     }
