+++ src/main/java/chess/domain/piece/movePattern/AbstractPawnMovePattern.java
@@ -1,6 +1,7 @@
-package chess.domain.piece;
+package chess.domain.piece.movePattern;
 
 import chess.domain.board.Position;
+import chess.domain.piece.Color;
 
 public abstract class AbstractPawnMovePattern implements MovePattern {
