+++ src/main/java/chess/domain/piece/movePattern/AbstractStraightMovePattern.java
@@ -1,8 +1,7 @@
-package chess.domain.piece;
+package chess.domain.piece.movePattern;
 
 import chess.domain.board.Direction;
 import chess.domain.board.Position;
-
 import java.util.List;
 
 public abstract class AbstractStraightMovePattern implements MovePattern {
