+++ src/main/java/chess/domain/piece/movePattern/BlackPawnMovePattern.java
@@ -1,9 +1,8 @@
-package chess.domain.piece;
+package chess.domain.piece.movePattern;
 
 import chess.domain.board.Direction;
 import chess.domain.board.Position;
 import chess.domain.board.Row;
-
 import java.util.List;
 
 public final class BlackPawnMovePattern extends AbstractPawnMovePattern {
