+++ src/main/java/chess/domain/state/Ready.java
@@ -3,7 +3,6 @@ package chess.domain.state;
 import chess.domain.board.Board;
 import chess.domain.board.Position;
 import chess.domain.piece.Piece;
-
 import java.util.Map;
 
 public abstract class Ready implements State {
