+++ src/main/java/chess/domain/state/State.java
@@ -4,7 +4,6 @@ import chess.domain.board.Position;
 import chess.domain.board.Result;
 import chess.domain.piece.Color;
 import chess.domain.piece.Piece;
-
 import java.util.Map;
 
 public interface State {
